Remove CompactionVectorValues -- it actually makes things worse since it forces retaining the original ByteBuffer on top of the float[] that we deserialize it to for the postings map.  With ConcurrentVectorValues we only add a bit of overhead on top of the float[] reference.
